requirements.txt
openpyxl
customtkinter
pillow
pandas
rapidfuzz
multiprocess
requests
aiohttp